% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_auxiliares.R
\name{astericos}
\alias{astericos}
\title{Gera uma sequência de estrelas para significância estatística}
\usage{
astericos(n, barra = T)
}
\arguments{
\item{n}{Um valor inteiro indicando o número de asteriscos desejados.}

\item{barra}{Um valor lógico. Se `TRUE`, inclui barras invertidas (`\\*`) nos asteriscos; se `FALSE`, usa apenas `*`. O padrão é `TRUE`.}
}
\value{
Uma string com a sequência de asteriscos formatada.
}
\description{
Esta função retorna uma sequência de asteriscos conforme o número especificado,
com opção de incluir barras invertidas.
}
\examples{
asteriscos(3) # "\\*\\*\\*"
asteriscos(2, barra = FALSE) # "**"
}
