% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva_continua.R
\name{desc_uni_continua}
\alias{desc_uni_continua}
\title{Descrição de Variáveis Contínuas}
\usage{
desc_uni_continua(
  vari,
  nome,
  texto = T,
  grafico = T,
  digitos = 2,
  idioma = "PT",
  virgula = F,
  estat_tab = "auto",
  nomes_tab = "auto",
  qqgraf = F,
  ...
)
}
\arguments{
\item{vari}{Vetor numérico com os dados da variável contínua.}

\item{nome}{Nome da variável (como string) para ser utilizado nos gráficos e no texto gerado.}

\item{texto}{Se TRUE, gera uma descrição textual dos resultados. Caso contrário, não gera texto (padrão é TRUE).}

\item{grafico}{Se TRUE, gera gráficos para visualizar a distribuição dos dados. Caso contrário, não gera gráficos (padrão é TRUE).}

\item{digitos}{Número de dígitos para arredondamento nas estatísticas (padrão é 2).}

\item{idioma}{Idioma para a geração do texto ('PT' ou 'EN'; o padrão é 'PT').}

\item{virgula}{Se TRUE, utiliza vírgula como separador decimal no texto e gráficos (padrão é FALSE).}

\item{bins}{Número de bins para o histograma (padrão é 20).}

\item{cor}{Cor a ser usada nos gráficos (padrão é 'cyan4').}
}
\value{
Retorna uma lista com os seguintes elementos:
\item{result}{Dataframe com as estatísticas descritivas da variável.}
\item{texto}{Texto interpretativo gerado, se o parâmetro texto for TRUE.}
\item{interp}{Resumo interpretativo em formato de texto resumido.}
\item{grafico}{Gráficos gerados, se o parâmetro grafico for TRUE.}
}
\description{
Esta função realiza uma descrição estatística de uma variável contínua, incluindo o cálculo de estatísticas básicas, como média, desvio padrão, coeficiente de variação, e teste de normalidade. Também gera um texto interpretativo e gráficos para visualizar a distribuição dos dados.
}
\examples{
# Criando um vetor de dados contínuos
dados <- rnorm(100)

# Chamando a função com texto e gráfico
resultado <- desc_uni_continua(dados, nome = "Variável X")

# Visualizando os resultados
resultado$grafico

}
