% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva_categorica.R
\name{grafico_categorica}
\alias{grafico_categorica}
\title{Gera gráfico de barras ou pizza para variáveis categóricas}
\usage{
grafico_categorica(
  var,
  nome,
  niveis = "auto",
  cor = "cyan4",
  ordenar = "auto",
  virgula = F,
  digitos = 1,
  forcarpizza = F,
  orient = "auto",
  nas = F
)
}
\arguments{
\item{var}{vetor (obrigatório) A variável categórica que será analisada.}

\item{nome}{string (obrigatório) O nome da variável, que será exibido nos títulos dos gráficos.}

\item{niveis}{vetor de strings (opcional, padrão = "auto") Os níveis da variável categórica.
Se definido como "auto", os níveis serão automaticamente extraídos da variável.}

\item{cor}{string (opcional, padrão = "cyan4") Cor das barras ou das fatias do gráfico.}

\item{ordenar}{Opções: T, F e 'auto' (opcional, padrão = 'auto') Se \code{TRUE}, as categorias serão ordenadas por
frequência. Caso contrário, elas são apresentadas na ordem original. Se 'auto', ordenará characters e não ordenará factors.}

\item{virgula}{lógico (opcional, padrão = FALSE) Se \code{TRUE}, utiliza vírgula como separador decimal nos rótulos dos gráficos (caso contrário, usa ponto).}

\item{digitos}{numérico (opcional, padrão = `1`) número de casas decimais no percentual nos labels dos gráficos.}

\item{forcarpizza}{lógico (opcional, padrão = FALSE) se devemos forçar um gráfico de pizza ou deixa ele escolher (pizza para dois níveis e barras para mais de dois)}

\item{orient}{opcional (padrão = "auto"). Se auto, opta por vertical se a quantidade de categorias for menor que 7 e o tamanho dos labels de todas categorias for menor que 25 e horizontal caso contrário. Outras opções são "v" (gráfico vertical) e "h" (gráfico horizontal)}

\item{nas}{Lógico. Se `TRUE`, inclui as frequências de valores ausentes (NA) no gráfico. Padrão: `FALSE`}
}
\description{
Esta função gera um gráfico de barras ou gráfico de pizza para variáveis categóricas,
com a possibilidade de personalizar a cor das barras e a ordem das categorias.
O gráfico exibe a frequência de cada nível da variável, bem como o percentual
correspondente a cada categoria.
}
\examples{
## Mais de dois níveis, ordenando
variavel = factor(c("A", "B", "C", "A", "A", "B", "C")) # Três níveis, ordenado do mais frequente para o menos frequente
grafico_categorica(var = variavel, nome = "Categoria", ordenar = TRUE)

## Mais de dois níveis, sem ordenar
variavel = factor(c("10 ou menos","11 a 20", "11 a 20","21 a 30"))
grafico_categorica(var = variavel, nome = "Faixa etária", ordenar = FALSE)
# Repare que é interessante não ordenar quando a variável for categórica ordinal, com objetivo de manter a ordem que faz sentido da variável

## Dois níveis
variavel = factor(c("Masculino", "Feminino", "Masculino"))
grafico_categorica(var = variavel, nome = "Gênero")

}
