% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva_categorica.R
\name{grafico_categorica_vert}
\alias{grafico_categorica_vert}
\title{Gera gráfico de barras horizontal para variáveis categóricas}
\usage{
grafico_categorica_vert(
  var,
  nome,
  niveis = "auto",
  cor = "cyan4",
  ordenar = T,
  virgula = F
)
}
\arguments{
\item{var}{vetor (obrigatório) A variável categórica que será analisada.}

\item{nome}{string (obrigatório) O nome da variável, que será exibido no título do gráfico.}

\item{niveis}{vetor de strings (opcional, padrão = "auto") Os níveis da variável categórica.
Se definido como "auto", os níveis serão automaticamente extraídos da variável.}

\item{cor}{string (opcional, padrão = "cyan4") Cor das barras no gráfico.}

\item{ordenar}{lógico (opcional, padrão = TRUE) Se \code{TRUE}, as categorias serão ordenadas
por frequência. Caso contrário, serão apresentadas na ordem definida pelo parâmetro \code{niveis}.}

\item{virgula}{lógico (opcional, padrão = FALSE) Se \code{TRUE}, utiliza vírgula como separador
decimal nos rótulos do gráfico (caso contrário, usa ponto).}
}
\value{
Um objeto \code{ggplot} contendo o gráfico de barras horizontal.
}
\description{
Esta função gera um gráfico de barras horizontal para variáveis categóricas,
apresentando a frequência relativa de cada nível da variável. O gráfico permite
personalização da cor, ordenação das categorias e definição dos níveis. Os rótulos
de cada barra exibem a frequência absoluta e o percentual correspondente.
}
\details{
Esta função utiliza o pacote \code{ggplot2} para gerar um gráfico de barras horizontal,
ideal para visualizar a distribuição de categorias de forma proporcional. A frequência
absoluta (\code{Freq}) e relativa (\code{Freq_rel}) são calculadas internamente e usadas
para compor o gráfico e os rótulos.
}
\examples{
## Variável categórica com mais de dois níveis, ordenando
variavel = factor(c("A", "B", "C", "A", "A", "B", "C"))
grafico_categorica_vert(var = variavel, nome = "Categorias", ordenar = TRUE)

## Variável categórica com ordem definida
variavel = factor(c("Idade: 0-10", "Idade: 11-20", "Idade: 11-20", "Idade: 21-30"))
niveis_definidos = c("Idade: 0-10", "Idade: 11-20", "Idade: 21-30")
grafico_categorica_vert(var = variavel, nome = "Faixa Etária", niveis = niveis_definidos, ordenar = FALSE)

}
