% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva_continua.R
\name{graficos_continua}
\alias{graficos_continua}
\title{Função para Gerar Gráficos de Variáveis Contínuas}
\usage{
graficos_continua(
  var,
  nome,
  tipo = "ambos",
  bins = 20,
  cor = "cyan4",
  digitos = 2,
  idioma = "PT",
  virgula = F,
  xmin = "auto",
  xmax = "auto",
  labels = T
)
}
\arguments{
\item{var}{Vetor numérico com os dados da variável contínua.}

\item{nome}{Nome da variável (como string) para ser utilizado nos gráficos.}

\item{tipo}{Quais gráficos serão plotados. Sendo o padrão "ambos" (ambos: Histograma e boxplot alinhados); "hist" histograma e "box" boxplot}

\item{bins}{Número de bins para o histograma (padrão é 20).}

\item{cor}{Cor a ser usada nos gráficos (padrão é 'cyan4').}

\item{digitos}{Número de dígitos para arredondamento nas estatísticas (padrão é 2).}

\item{idioma}{Idioma para a geração do texto ('PT' ou 'EN'; o padrão é 'PT').}

\item{virgula}{Se TRUE, utiliza vírgula como separador decimal no texto e gráficos (padrão é FALSE).}

\item{xmin}{Valor mínimo do eixo x do gráfico (padrão é 'auto').}

\item{xmax}{Valor máximo do eixo x do gráfico (padrão é 'auto').}
}
\value{
Retorna um gráfico combinando um boxplot e um histograma com a distribuição dos dados.
}
\description{
Esta função gera gráficos de boxplot e histograma para visualizar a distribuição de uma variável contínua.
}
\examples{
dados <- rnorm(100)
graficos_continua(dados, nome = "Variável X")

}
