% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_auxiliares.R
\name{magic_format}
\alias{magic_format}
\title{Formata números com casas decimais personalizadas}
\usage{
magic_format(numero, dig = 2, virgula = F)
}
\arguments{
\item{numero}{Um valor numérico ou vetor a ser formatado.}

\item{dig}{Um valor inteiro indicando o número de casas decimais. O padrão é `2`.}

\item{virgula}{Um valor lógico. Se `TRUE`, utiliza vírgula como separador decimal; se `FALSE`, utiliza ponto. O padrão é `FALSE`.}
}
\value{
Uma string representando o número formatado com o separador decimal definido.
}
\description{
Esta função formata números com a quantidade desejada de casas decimais,
permitindo configurar o separador decimal. Apenas repete os caracteres não numéricos do vetor.
}
\examples{
magic_format(3.14159) # "3.14"
magic_format(3.14159, dig = 3) # "3.142"
magic_format(3.14159, virgula = TRUE) # "3,14"



}
