% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva_continua.R
\name{magic_qqplot}
\alias{magic_qqplot}
\title{Gera um gráfico Q-Q plot com teste de normalidade}
\usage{
magic_qqplot(x, nome, virgula = F, digitos = 2)
}
\arguments{
\item{x}{Vetor numérico com os dados da amostra.}

\item{nome}{Uma string que representa o nome da variável a ser exibida no título do gráfico.}

\item{virgula}{Um valor lógico. Se `TRUE`, utiliza vírgula como separador decimal nas anotações; se `FALSE`, utiliza ponto. O padrão é `FALSE`.}

\item{digitos}{Um valor inteiro indicando o número de casas decimais a ser utilizado na formatação do teste de normalidade. O padrão é `2`.}
}
\value{
Um objeto gráfico gerado pelo pacote `ggpubr`, contendo o Q-Q plot com anotações do teste de normalidade.
}
\description{
Esta função cria um gráfico de probabilidade Q-Q (Q-Q plot) para uma amostra numérica,
adicionando uma anotação com os resultados do teste de Shapiro-Wilk de normalidade.
}
\examples{
set.seed(123)
dados <- rnorm(100)
magic_qqplot(dados, nome = "Amostra de teste")

magic_qqplot(dados, nome = "Exemplo", virgula = TRUE, digitos = 3)
}
