% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_auxiliares.R
\name{ponto_para_virgula}
\alias{ponto_para_virgula}
\title{Substitui ponto por vírgula em um vetor}
\usage{
ponto_para_virgula(vetor, virgula = F)
}
\arguments{
\item{vetor}{Um vetor de strings onde os pontos serão substituídos por vírgulas, caso o parâmetro `virgula` seja `TRUE`.}

\item{virgula}{Um valor lógico que determina se a substituição de ponto por vírgula deve ocorrer. O valor padrão é `FALSE` (não substitui).}
}
\value{
O vetor original com os pontos substituídos por vírgulas (se `virgula` for `TRUE`).
}
\description{
Esta função substitui todos os pontos (".") por vírgulas (",") em um vetor de strings, com base no parâmetro `virgula`.
}
\keyword{internal}
