% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_auxiliares.R
\name{printvetor}
\alias{printvetor}
\title{printvetor}
\usage{
printvetor(vetor, idioma = "PT", aspas = TRUE)
}
\arguments{
\item{vetor}{Vetor de caracteres a ser impresso. Os elementos serão unidos
através de uma conjunção.}

\item{idioma}{Idioma utilizado na conjunção entre os elementos. Pode ser
"PT" para português (padrão) ou "EN" para inglês.}

\item{aspas}{Lógico. Se `TRUE`, coloca aspas ao redor dos elementos. O valor
padrão é `TRUE`.}
}
\value{
Uma string com os elementos do vetor unidos com a conjunção apropriada, e com
ou sem aspas, dependendo do valor do parâmetro \code{aspas}.
}
\description{
Esta função formata e retorna um vetor como uma string, unindo seus elementos
com uma conjunção apropriada, dependendo do idioma (Português ou Inglês). A função
também permite adicionar aspas ao redor dos elementos do vetor, caso desejado.
}
\details{
A função \code{printvetor} é útil para gerar listagens formatadas de itens,
com a possibilidade de ajustar a conjunção ("e" ou "and") e a formatação
(com ou sem aspas) dos elementos do vetor. Quando o vetor contém apenas um
elemento, ele é retornado sem a conjunção.
}
\examples{
vetor_exemplo <- c("maçã", "banana", "laranja")

# Imprimindo os elementos com aspas e em português
printvetor(vetor_exemplo, idioma = "PT", aspas = TRUE)

# Imprimindo os elementos sem aspas e em inglês
printvetor(vetor_exemplo, idioma = "EN", aspas = FALSE)

}
\keyword{internal}
