% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_auxiliares.R
\name{relatorio}
\alias{relatorio}
\title{Gera um relatório formatado a partir de uma lista.}
\usage{
relatorio(a, dig = 2)
}
\arguments{
\item{a}{Um vetor ou lista contendo os objetos a serem relatados. Pode incluir data.frames, listas, vetores de caracteres ou outros tipos de objetos.}

\item{dig}{Um número inteiro (default = 2). Define o número de casas decimais a serem exibidas nas tabelas de data.frames.}
}
\description{
Esta função percorre uma lista de objetos e gera um relatório de acordo com o tipo de cada item. Para data.frames, a função imprime uma tabela formatada (com kable). Para listas, chama a própria função recursivamente. Para vetores de caracteres, imprime os itens no formato de texto (com cat), e para outros objetos, apenas imprime o valor.
}
\details{
A função utiliza o pacote `knitr` para formatar as tabelas quando encontra objetos do tipo `data.frame`. Quando encontra listas, a função é chamada recursivamente para gerar um relatório para cada item. Para objetos do tipo `character`, os valores são impressos em formato de texto.
}
\examples{
# Exemplo de uso:
lista_exemplo <- list("",
  data.frame(A = 1:3, B = c(4.567, 5.678, 6.789)),
  "Texto de exemplo",
  list(data.frame(C = 1:2, D = c(7.890, 8.901)))
)
relatorio(lista_exemplo)

}
