% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_auxiliares.R
\name{theme_icic}
\alias{theme_icic}
\title{Tema customizado para gráficos do ggplot2}
\usage{
theme_icic(sentido = "v")
}
\arguments{
\item{sentido}{string (opcional, padrão = "v") Determina a orientação do gráfico:
\code{"v"} para gráficos verticais e \code{"h"} para gráficos horizontais.}
}
\value{
Um objeto \code{theme} do pacote \code{ggplot2} que pode ser aplicado diretamente
a um gráfico \code{ggplot}.
}
\description{
Esta função define um tema customizado baseado no \code{ggthemes::theme_clean()},
com ajustes específicos para facilitar a visualização de gráficos horizontais
ou verticais. O tema modifica o fundo, as grades e o alinhamento de textos no gráfico.
}
\details{
O tema \code{theme_icic} é projetado para melhorar a apresentação de gráficos
criados com o pacote \code{ggplot2}. Ele inclui:
- Fundo transparente para o painel e o gráfico.
- Título e subtítulo centralizados.
- Ajustes nas grades principais:
  - Para gráficos verticais (\code{sentido = "v"}): grades verticais com linhas pontilhadas.
  - Para gráficos horizontais (\code{sentido = "h"}): grades horizontais com linhas pontilhadas.
}
\examples{
library(ggplot2)

dados = rnorm(100,50,10)

# Aplicando o tema em gráficos verticais
p <- ggplot(NULL, aes(x=dados)) +
     geom_histogram()

p1 = p + theme_icic(sentido = "v")
print(p1)

# Aplicando o tema em gráficos horizontais
p <- ggplot(NULL, aes(x=dados)) +
     geom_histogram()

p2 = p + theme_icic(sentido = "y")
print(p2)
}
\keyword{internal}
