% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_auxiliares.R
\name{vetor_comsep_c}
\alias{vetor_comsep_c}
\title{Função para separar um vetor em partes com base em um corte inicial}
\usage{
vetor_comsep_c(vec, corteinicial)
}
\arguments{
\item{vec}{Vetor de strings a ser dividido. Cada elemento do vetor será tratado separadamente.}

\item{corteinicial}{Valor numérico indicando a posição inicial para o corte do texto.}
}
\value{
Retorna um vetor de strings com as partes separadas por cortes. Caso o texto não possa ser dividido, retorna `NA`.
}
\description{
Essa função divide cada elemento de um vetor de texto em substrings, utilizando um corte inicial e espaços como delimitadores. É útil para formatar ou dividir texto de maneira estruturada.
}
\keyword{internal}
