% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_auxiliares.R
\name{pvalor}
\alias{pvalor}
\title{Formata p-valores com estrelas}
\usage{
pvalor(p)
}
\arguments{
\item{p}{Um valor numérico representando o p-valor.}
}
\value{
Uma string formatada do p-valor com as estrelas correspondentes aos níveis de significância.
}
\description{
Essa função é usada internamente para formatar p-valores de forma consistente,
adicionando estrelas conforme os níveis de significância estatística.
}
\keyword{internal}
