% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva_categorica.R
\name{quiqua_aderencia}
\alias{quiqua_aderencia}
\title{quiqua_aderencia}
\usage{
quiqua_aderencia(
  vetor,
  nomecat,
  niveis = "auto",
  ordenar = T,
  dig = 2,
  virgula = F
)
}
\arguments{
\item{vetor}{Vetor ou lista contendo a variável categórica a ser analisada.}

\item{nomecat}{Nome da variável categórica para uso nos textos descritivos.}

\item{niveis}{Ordem dos níveis da variável categórica. Padrão é `"auto"`, que utiliza os níveis da variável.}

\item{ordenar}{Lógico. Se `TRUE`, os resultados são ordenados por resíduos padronizados. Padrão: `TRUE`.}

\item{dig}{Número de casas decimais para arredondamento. Padrão: `2`.}

\item{virgula}{Lógico. Se `TRUE`, utiliza vírgula como separador decimal. Padrão: `FALSE`.}
}
\value{
Um objeto `list` com os seguintes elementos:
\item{texto}{Texto interpretativo do teste qui-quadrado e dos resíduos.}
\item{tabela}{Dataframe contendo a proporção observada de cada categoria com seu intervalo de confiança, os resíduos padronizados e significância correspondente considerando a correção de bonferroni. Resíduos significativos são assinalados com `*`.}
}
\description{
Executa o teste qui-quadrado de aderência para verificar se as frequências observadas em uma variável categórica diferem significativamente de frequências esperadas iguais entre as categorias. Também calcula os resíduos padronizados e identifica categorias com frequências significativamente maiores ou menores.
}
\note{
- Quando o número de categorias é grande, o teste pode perder poder devido ao ajuste para comparações múltiplas.
- As interpretações do tamanho do efeito baseiam-se no V de Cramer e variam com o número de graus de liberdade.
- Caso a variável tenha apenas uma categoria, a função retorna uma mensagem informando a impossibilidade de realizar o teste.
- As funções `chisq.test` e `prop.test` são usadas para o cálculo do teste qui-quadrado e dos intervalos de confiança, respectivamente.
}
\examples{
# Dados fictícios
dados <- sample(c("A", "B", "C"), size = 50, replace = TRUE, prob = c(0.1, 0.5, 0.4))

# Rodando o teste qui-quadrado de aderência
resultado <- quiqua_aderencia(vetor = dados, nomecat = "Categorias")

# Gerando o relatório
relatorio(resultado,pularprimeiro=F)

}
